backend/requirements.txt
openai>=1.0.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn>=0.30.0
